+++ ui/src/pages/LinePage.vue
@@ -17,18 +17,18 @@ const defaultOptions = computed((): PredefinedGraphOption<'scatterplot'>[] | nul
       && p.spec.annotations?.['pl7.app/vdj/isScore'] === undefined
     ))].spec;
 
-  const frequencyColSpec = getColSpec('pl7.app/vdj/frequency');
+  const frequencyColSpec = getColSpec('pl7.app/frequency');
   const defaults: PredefinedGraphOption<'scatterplot'>[] = [
     {
       inputName: 'y',
       selectedSource: frequencyColSpec,
     },
-    // pl7.app/vdj/condition
+    // pl7.app/condition
     {
       inputName: 'x',
       selectedSource: frequencyColSpec.axesSpec[1],
     },
-    // pl7.app/vdj/clonotypeKey
+    // input element axis (vdj clonotypeKey / peptide variantKey)
     {
       inputName: 'grouping',
       selectedSource: frequencyColSpec.axesSpec[0],
@@ -57,8 +57,8 @@ const dataColumnPredicate = (spec: PColumnSpec) =>
   inputElementAxis.value !== undefined
   && spec.axesSpec.length === 2
   && spec.axesSpec[0].name === inputElementAxis.value
-  && spec.axesSpec[1].name === 'pl7.app/vdj/condition'
-  && (spec.name === 'pl7.app/vdj/frequency' || spec.name === 'pl7.app/vdj/enrichmentVsBaseline');
+  && spec.axesSpec[1].name === 'pl7.app/condition'
+  && (spec.name === 'pl7.app/frequency' || spec.name === 'pl7.app/enrichmentVsBaseline');
 
 const metaColumnPredicate = (spec: PColumnSpec) =>
   inputElementAxis.value !== undefined