+++ ui/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @platforma-open/milaboratories.clonotype-enrichment.ui
 
+## 4.0.0
+
+### Major Changes
+
+- d74e7e0: Support peptides
+
+### Patch Changes
+
+- Updated dependencies [d74e7e0]
+  - @platforma-open/milaboratories.clonotype-enrichment.model@4.0.0
+
 ## 3.2.6
 
 ### Patch Changes